Updated admin products

+++ products/models.py
@@ -51,7 +51,7 @@ class Category(models.Model):
 
 class Product(models.Model):
     category = models.ForeignKey('Category', null=True, blank=True, on_delete=models.SET_NULL)
-    number_id = models.DecimalField(max_digits=6, decimal_places=2, null=True, blank=True)
+    number_id = models.IntegerField(null=True, blank=True)
     name = models.CharField(max_length=254)
     brand = models.CharField(max_length=254, null=True, blank=True)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
